Refactored configuration code
