+++ jmx_prometheus_common/src/main/java/io/prometheus/jmx/common/ConfigurationException.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.prometheus.jmx.common.http;
+package io.prometheus.jmx.common;
 
 import java.util.function.Supplier;
