+++ jmx_prometheus_common/src/main/java/io/prometheus/jmx/common/authenticator/Credentials.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.prometheus.jmx.common.http.authenticator;
+package io.prometheus.jmx.common.authenticator;
 
 import java.util.Objects;
