+++ jmx_prometheus_common/src/test/java/io/prometheus/jmx/common/http/authenticator/CredentialsCacheTest.java
@@ -18,6 +18,8 @@ package io.prometheus.jmx.common.http.authenticator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import io.prometheus.jmx.common.authenticator.Credentials;
+import io.prometheus.jmx.common.authenticator.CredentialsCache;
 import java.nio.charset.StandardCharsets;
 import org.junit.Test;
