+++ jmx_prometheus_common/src/test/java/io/prometheus/jmx/common/http/authenticator/CredentialsTest.java
@@ -19,6 +19,7 @@ package io.prometheus.jmx.common.http.authenticator;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 
+import io.prometheus.jmx.common.authenticator.Credentials;
 import org.junit.Test;
 
 public class CredentialsTest {
