+++ jmx_prometheus_common/src/test/resources/MapAccessorTest.yaml
